Use struct to implement Policy
This refactoring PR aims to leverage OOP principles to implement Policy and PolicyEngine.
Policy is an interface that has four methods:

* Name() which returns the name of policy
* Metadata() which returns the metadata related to the policy
* Help() which returns the help text
* Validate() which takes an image tag and validates whether the image satisfies certain criteria

Each policy is a type (i.e construct) that must implement above functions.

The PolicyRunner takes a slice of Policy types and run the Validate() method on each policy with the given Image ID and aggregates the result.