Make the runs_as_nonroot test more robust
If there was any output on stderr, this check would usually fail. This
patch makes it a bit more robust by separating stdout and stderr, and
just relying on the stdout stream. It also uses TrimSpace instead of
TrimSuffix, as the suffix wasn't always CRLF, but sometime just LF,
and this would also cause the check to fail.

Signed-off-by: Brad P. Crochet <brad@redhat.com>