Remove logger as a parameter
Originally, the logger was being passed around, and was a part of the
main interfaces. This removes that dependency, and allows code to just
import logrus as 'log' and be able to log without setup or passing the
logger around. This patch maintains backwards compatibility with the
functionality that existed for the logger.

Signed-off-by: Brad P. Crochet <brad@redhat.com>