Extract common test logic into helper functions
The unit tests for each of the checks follow a basic pattern.

Set up the fake engine
Verify that is passes Validate
Modify the fake engine for a failure mode
Verify that it does not pass Validate
Verify that Podman errors are handled properly

This extracts the verifications into helper functions that can be reused
to cut down on the boilerplate.

Fixes #44