Add a new command to run preflight in an unshare namespace
Introduces a new command: 'preflight check run'. This command will inspect
the environment for specific values, and exit if they are not found.
Values are:
PREFLIGHT_EXEC_CHECK - The name of a check as given by check.Name()
PREFLIGHT_EXEC_IMAGE - The image under test

What this gives us is the ability to run a test within a podman
namespace, via the unshare command.

This will allow a container to be created, mounted, have the filesystem
directly operated upon, then unmounted and removed, all within the context
of podman unshare.

HasNoProhibitedPackages has been modified to utilize this behavior, in order
to operator on the RPM db directly, without the need for the 'rpm' executable,
which happens to not be available in either a scratch or microubi-based
image.

Fixes #104

Signed-off-by: Brad P. Crochet <brad@redhat.com>