Make sure logs are written to the artifacts/ dir
In order to make sure that logs are persisted in a containerized
run, this writes the logs to a relative dir of 'artifacts/'.

It is assumed that the workdir for the container is '/'

Also, multiple invocations of unshare will overwrite the logfile
of a previous run. This will separate them out.

Signed-off-by: Brad P. Crochet <brad@redhat.com>