Remove unwanted log message
A log message, at default "info" level (since it was a log.Println)
was in use when first debugging the Untar function, but it was an
extremely verbose log, so even with trace it is overwhelming, and
doesn't add any good info for the end user. So, this removes it.

Signed-off-by: Brad P. Crochet <brad@redhat.com>