Reorganize code for clarity
Some of the code was put (likely prematurely) into a "utils" package, even
if just a single package was using it. The moves that "utility" code closer
to the consumer, and in some cases, no longer exported.

This also moves some of the truly utility functions into a more sane package
(such as artifacts) that are shared, but make the package name more
descriptive, and therby allowing a shortening of the function name itself.

Signed-off-by: Brad P. Crochet <brad@redhat.com>