Fix subtle bug that was not completing operator-sdk calls
If FATA exists in stderr when calling operator-sdk, it means there
was an error in the execution of the command, not in the check being
run. This detects that and marks as failed instead of error
appropriately. It has the side effect of actually showing the error
in the logs.

Fixes #328

Signed-off-by: Brad P. Crochet <brad@redhat.com>