Increase the default wait time for scorecard tests
* Fixes #329

This PR introduces a new field for the scorecard runtime configuration.
Scorecard requires pulling the busybox image (among others). The default
wait time of 30s is not enough to pull the dependencies and run the check.
This PR increases the wait time to 120s.

Found a relevant issue on github:
https://github.com/operator-framework/operator-sdk/issues/4148