Modify runtime assets to dynamically pull the digest
Instead of hard-coding the digest, this patch now pulls the digest dynamically
at runtime, when 'preflight runtime-assets' is run. The output would then be
used to mirror to a disconnected registry. Also, when preflight is run in a
disconnected cluster, this digest can be passed via PFLT_SCORECARD_IMAGE.

This predicates that the preflight runtime-assets be run on a connected host.
But because the image can be passed by envvar, preflight check operator can
be run on a disconnected cluster without having the sha hard-coded.

Signed-off-by: Brad P. Crochet <brad@redhat.com>