Introduce Pyxis Engine
This is the start of the Pyxis engine. It also adds support for a
PYXIS_HOST envvar, and a PYXIS_API_TOKEN envvar and CLI param.

The PyxisEngine should export useful methods to those that use it.
It is up to the user to create interfaces if necessary for testing
purposes.

Signed-off-by: Brad P. Crochet <brad@redhat.com>