Parameterize the IMAGE_REGISTRY in Actions
In order to make testing of actions more useful, this patch sets
the image registry based on the IMAGE_REGISTRY secret on the
current repository.

It also triggers on the 'ghactions' branch, which facilitates
using that branch to test changes to the CI without having to
modify the main repo.

Signed-off-by: Brad P. Crochet <brad@redhat.com>