Limit the usage of log.Info
Since log.Info is going to be the default log level, we want to make sure
that it is not over-used. Most of the things in the internal package should
be debug anyway.

This just makes the default Info level have useful information that we want
all users to see, not debug-type messages.

Signed-off-by: Brad P. Crochet <brad@redhat.com>