Pass docker config to the results upload
The docker config was not being passed to the results. This adss them
to the results, and makes the env var the same as the check operator
command.

This also rearranges some of the initialization code so that the viper
configs are initialized in time to be able to shortcut whether
--docker-config is required or not. If the PFLT_DOCKERCONFIG envvar
is set, it does not require the flag.

Signed-off-by: Brad P. Crochet <brad@redhat.com>