Introduce context param throughout the codebase
Context should be shared from the highest level down through all
calls. This is the beginning of making that a reality. Once context
is being passed, we can start to add things to the context, like
params from viper, so that the parts of the codebase that do not
have to do with the CLI don't know anything about viper. Consider
it a separation of concerns.

Signed-off-by: Brad P. Crochet <brad@redhat.com>