Fix source rpm name parsing
Some source rpm names are breaking the parsing logic we have (that was
very close to the CVP version). We couldn't use the CVP version, as the
golang regex engine does not support positive lookaheads. So, this patch
instead utilizes Split on '-', then puts the source name back together
with any '-', minus the last two entries. Unit tests provided.

Signed-off-by: Brad P. Crochet <brad@redhat.com>