Utilize built-in cache of crane library
It was discovered that the HasModifiedFiles check was having to re-
pull the image under test. With this simple patch, we now utilize
the caching mechanism of crane, which has reduced the runtime of
HasModifiedFiles by an order of magnitude. It ends up being a
transparent change to the entire codebase, as the v1.Image passed
around in the ImageRef is an instance of the cached Image.

Fixes #533

Signed-off-by: Brad P. Crochet <brad@redhat.com>