Use graphql for CheckRedHatLayers
It was pointed out that if too many layer hashes were passed in the
URL for a REST request, it could cause a 500 for the length of the
URL being too long. This switches the CheckRedHatLayers pyxis query
to use the GraphQL endpoint, and send the query in the body, which
has no such limit.

Signed-off-by: Brad P. Crochet <brad@redhat.com>