Make the pulling of the project to check for scratch optional
In order to facilitate the running of preflight without the requirement
of a certification ID and API key, we will check if the certification
ID is provided. If it is not, we will skip checking the project before
running the checks. If --submit is given, the id and API key will become
required again.

Refers to #552
Fixes #549

Signed-off-by: Brad P. Crochet <brad@redhat.com>