Make the Connect URLs match the environment given
The major change here is that if a user provides a Pyxis host directly
(which should not be usual) then the URL returned for connect will
just be prod.

Also refactor tests a little bit to make them more readable. Was already
in the file adding tests for this patch so made them more consistent.

Fixes #576

Signed-off-by: Brad P. Crochet <brad@redhat.com>