Backports from main (#590)
* updating check container check url in all tests to point to official 4.9 cert docs (#581)
* updating check container recipe to add in support for private registries (#583)
* adding in PFLT_CERTIFICATION_PROJECT_ID to CONFIG info
* Switch to a go-release-action for building binaries

For now, this uses a fork off the action on GitHub. It adds support
for leaving the binaries uncompressed. A PR has been opened on the
upstream, and we will switch to that once that lands (assumng it
does).

This patch also changes the trigger to "on release published", and
removes the "push to tags or branch ghactions".

Fixes #578

Signed-off-by: Brad P. Crochet <brad@redhat.com>
Co-authored-by: Adam D. Cornett <adc@redhat.com>