Fix a data race exposed by race checker
When adding the -race flag to go test along with -cover, a couple of
data races were exposed by the tool. This fixes those data races, and
introduces 'make cover' as a target in the Makefile.

The wg.Wait() was being executed before anything was added to the
WaitGroup. Therefore, it was possible that the channel might close
before any work was done. This makes sure that the WaitGroup has
something to wait on before calling Wait.

Signed-off-by: Brad P. Crochet <brad@redhat.com>