Fix a panic in DeployableByOLM
When the OpenShift client was refactored, a subtle timing bug was
introduced. The openshift.GetCsv function was now returning a nil
for the CSV when not found. The csvStatusSucceeded function was
then checking for a value in the CSV, but not checking whether it
was nil or not. If it happened to be called before the CSV was at
least partially established, it would panic. This should take care
of that.

Signed-off-by: Brad P. Crochet <brad@redhat.com>