Do not write artifacts to current dir in tests
A lot of the tests would create artifacts folders in the directory
under test. They would also create preflight.log in those same
locations. This makes sure that that does not occur.

It also removes the side effect from the artifacts.Path() call.
That call would create the artifacts dir, and that's just very
unexpected. Instead, this moves the dir creation side effect into
the artifacts.WriteFile() call, which is much more an expected
behaviour.

Signed-off-by: Brad P. Crochet <brad@redhat.com>