Fix some tests that would fail with existing env vars
Some of the unit tests would fail if KUBECONFIG or PFLT_INDEXIMAGE
were already present in the environment. This patch cleans those up
a bit by checking for the existence, clearing them when required, and
cleaning up appropriately after the test is run.

This also includes minor renaming of some of the test suites, as they
overlapped, and it was confusing where the error was actually coming
from.

Signed-off-by: Brad P. Crochet <brad@redhat.com>