Grab the image that matches the arch of the binary
Crane defaults to "linux/amd64" as the platform, unless an option
is passed with WithPlatform. This defaults to an OS of "linux", and
an architecture of runtime.GOARCH, which is basically set at build
time.

Additional work will be required to process all arches of an image
in one go.

Related #807

Signed-off-by: Brad P. Crochet <brad@redhat.com>