Add insecure option to support self-signed and HTTP registries
Without this option, crane would always try to verify the TLS
certificates of the registry. This would not allow for self-signed
certificates, or registries that only use HTTP. In order to support
the workflow to start a registry, push to that local registry, and
then pull preflight (think CI system), this option is necessary.

This is considered the "official" method for building a local image
and testing against it.

Fixes #593

Signed-off-by: Brad P. Crochet <brad@redhat.com>