Ignore layer files that have security capability attrs
If a file in a layer has had setcap used on it, it will show up in
the layer, but be unchanged. However, there will be a PAX record
that shows the extended attrs. For now, preflight will ignore these
files, unless/until we find a situation where we should be more robust
in this check and possibly fail it.

Fixes #969

Signed-off-by: Brad P. Crochet <brad@redhat.com>