Clarify docker-config is required for public Dockerhub registries
While submitting images to Red Hat, I noticed we weren't able to pull even though the images are in a public DockerHub registry. Red Hat support explained that the `--docker-config` flag is required for DockerHub registries. Updating here to help others, as I assumed it wasn't needed since this was a public registry.