Only check first page of tags in `HasUniqueTag`
An image passes the check if it has more than one tag (`latest` can be
one of them) or if it has only one tag which is not `latest`.

We can validate these scenarios using just the first page of tags
returned from the registry, rather than requesting all tags for a
repository which may be a lot more.

Signed-off-by: Caleb Xu <caxu@redhat.com>